requirements.txt
streamlit==1.32.0
torch==2.2.0
torchvision==0.17.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.3.2
Pillow==10.2.0
matplotlib==3.8.3
altair==5.2.0
seaborn==0.13.2
nltk==3.8.1
kagglehub==0.3.12